+++ src/main/java/org/osiam/resources/scim/User.java
@@ -1177,6 +1177,10 @@ public final class User extends Resource implements Serializable {
             return this;
         }
 
+        /**
+         * @deprecated You should not need to set the meta attribute with a client. Will be removed in 1.12 or 2.0.
+         */
+        @Deprecated
         @Override
         public Builder setMeta(Meta meta) {
             super.setMeta(meta);